Separate AbstractDataTable into another file.
